refactor: update PostDTO import paths
