+++ src/constants/mockData.ts
@@ -1,4 +1,4 @@
-import { PostDTO } from "@/api/post";
+import { PostDTO } from "@/types/post";
 
 export const mockGuestbookItems: PostDTO[] = [
     {
